bad commits were reverted. We now have a valid and working version with the official 80211abg channels
git-svn-id: svn://scm.gforge.inria.fr/svnroot/cmac/trunk@92 82507c83-fcc2-4493-92f8-b30dc6791bf2

+++ models/cmac_tools.ex.c
@@ -17,8 +17,8 @@
 
 
 //list of channels and their associated frequency
-double channel_to_freq(short channel){
-	FIN(channel_to_freq(short channel));
+double channel_to_freq(int channel){
+	FIN(channel_to_freq(int channel));
 
 	//802.11bg
 	switch(channel){
